Integration with NERDTree plugin
Trigger the project script when the project directory is open in any
NERDTree window.

Use :NERDTreeCWD instead of :NERDTreeToggle in the #FindLoad() function.

Add g:ProjectConfig_NERDTreeIntegration option to skip NERDTree support
when set to v:true. Change uses of 0 and !0 as boolean values with uses
of v:false and v:true.

Create s:AddVimListeners() function so autocommands are only registered
if the plugin is used (in ~/.vimrc).

Add debugging function to expose script variables.

Fix recursion when executing :lchdir to source the project script.

Fix condition in #SetScript() checking for "~/vimfiles" on Windows.